api/requirements_fastapi.txt
gunicorn==22.0.0
uvicorn==0.23.2
fastapi==0.110.3
aiohttp==3.11.13
pydantic==2.10.6


pymongo==4.8.0
motor==3.5.1
slowapi==0.1.9
redis==5.0.7
python-dotenv==1.0.1
